plotting benchmarks moved to gui_test
